🔨 refactor: File name changed from utils.rs to parsers.rs
